Support arb. list filtering; memoize lists
