+++ ui/src/components/List/Card.jsx
@@ -1,0 +1,12 @@
+import React, { useContext } from "react";
+import { VisibilityContext } from "react-horizontal-scrolling-menu";
+
+export const Card = ({ onClick, selected, title, itemId, children }) => {
+  const visibility = useContext(VisibilityContext);
+
+  return (
+    <div onClick={() => onClick(visibility)} tabIndex={0}>
+      {children}
+    </div>
+  );
+};
